(split)bug 39039 delete dangling folders only when the default view is calendar or tasks.
http://bugzilla.zimbra.com/show_bug.cgi?id=39039

Copied from Perforce
 Change: 164154

+++ src/java/com/zimbra/cs/datasource/CalDavDataImport.java
@@ -272,6 +272,9 @@ public class CalDavDataImport extends MailItemImport {
 			int[] fidArray = new int[allFolders.size()];
 			int i = 0;
 			for (DataSourceItem f : allFolders.values()) {
+				Folder folder = mbox.getFolderById(octxt, f.itemId);
+				if (folder != null && folder.getDefaultView() != MailItem.TYPE_APPOINTMENT && folder.getDefaultView() != MailItem.TYPE_TASK)
+					continue;
 				fids.add(f.itemId);
 				fidArray[i++] = f.itemId;
 				DbDataSource.deleteAllMappingsInFolder(ds, f.itemId);
